CAMEL-16400: split unit and integration tests for camel-google-bigquery
