+++ components/camel-google/camel-google-bigquery/src/test/java/org/apache/camel/component/google/bigquery/integration/AsyncIT.java
@@ -14,32 +14,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.google.bigquery.integration.sql;
+package org.apache.camel.component.google.bigquery.integration;
 
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
 import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
+import org.apache.camel.ExchangePattern;
+import org.apache.camel.LoggingLevel;
 import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.bigquery.integration.BigQueryTestSupport;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.support.DefaultExchange;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIf;
 
-public class SqlQueryTest extends BigQueryTestSupport {
-    private static final String TABLE_ID = "test_sql_table";
+@EnabledIf(value = "org.apache.camel.component.google.bigquery.integration.BigQueryITSupport#hasCredentials",
+           disabledReason = "Credentials were not provided")
+public class AsyncIT extends BigQueryITSupport {
+    private static final String TABLE_ID = "asynctest";
 
     @EndpointInject("direct:in")
     private Endpoint directIn;
 
-    @EndpointInject("google-bigquery-sql:{{project.id}}: insert into {{bigquery.datasetId}}." + TABLE_ID
-                    + "(col1, col2) values (@col1, @col2)")
+    @EndpointInject("google-bigquery:{{project.id}}:{{bigquery.datasetId}}:" + TABLE_ID)
     private Endpoint bigqueryEndpoint;
 
     @EndpointInject("mock:sendResult")
@@ -58,30 +63,38 @@ public class SqlQueryTest extends BigQueryTestSupport {
         return new RouteBuilder() {
             public void configure() {
                 from(directIn)
-                        .routeId("InsertRow")
-                        .to(bigqueryEndpoint)
+                        .to("seda:seda");
+                from("seda:seda")
+                        .routeId("Async")
+                        .to(ExchangePattern.InOnly, bigqueryEndpoint)
+                        .log(LoggingLevel.INFO, "To sendresult")
                         .to(sendResult);
             }
         };
     }
 
     @Test
-    public void insertRecordBySql() throws Exception {
-        Exchange exchange = new DefaultExchange(context);
-        String uuidCol1 = UUID.randomUUID().toString();
-        String uuidCol2 = UUID.randomUUID().toString();
+    public void sendAsync() throws Exception {
+        List<Map> objects = new ArrayList<>();
+        for (int i = 0; i < 5; i++) {
+            Exchange exchange = new DefaultExchange(context);
+            String uuidCol1 = UUID.randomUUID().toString();
+            String uuidCol2 = UUID.randomUUID().toString();
 
-        Map<String, String> object = new HashMap<>();
-        object.put("col1", uuidCol1);
-        object.put("col2", uuidCol2);
-        exchange.getIn().setBody(object);
+            Map<String, String> object = new HashMap<>();
+            object.put("col1", uuidCol1);
+            object.put("col2", uuidCol2);
+            objects.add(object);
+            exchange.getIn().setBody(object);
+            producer.send(exchange);
+        }
+        sendResult.expectedMessageCount(5);
 
-        sendResult.expectedMessageCount(1);
-        sendResult.expectedBodiesReceived(1);
-        producer.send(exchange);
         sendResult.assertIsSatisfied(4000);
 
-        assertRowExist(TABLE_ID, object);
+        for (Map object : objects) {
+            assertRowExist(TABLE_ID, object);
+        }
     }
 
 }
